Rename: Separate posts and user API
